+++ workos/types/mfa/authentication_challenge_verification_response.py
@@ -1,0 +1,9 @@
+from workos.resources.workos_model import WorkOSModel
+from workos.types.mfa.authentication_challenge import AuthenticationChallenge
+
+
+class AuthenticationChallengeVerificationResponse(WorkOSModel):
+    """Representation of a WorkOS MFA Challenge Verification Response."""
+
+    challenge: AuthenticationChallenge
+    valid: bool
