+++ workos/types/mfa/authentication_factor_totp_and_challenge_response.py
@@ -1,0 +1,10 @@
+from workos.resources.workos_model import WorkOSModel
+from workos.types.mfa.authentication_challenge import AuthenticationChallenge
+from workos.types.mfa.authentication_factor import AuthenticationFactorTotpExtended
+
+
+class AuthenticationFactorTotpAndChallengeResponse(WorkOSModel):
+    """Representation of an authentication factor and authentication challenge response as returned by WorkOS through User Management features."""
+
+    authentication_factor: AuthenticationFactorTotpExtended
+    authentication_challenge: AuthenticationChallenge
